+++ src/modules/mod.rs
@@ -1,0 +1,4 @@
+pub mod soe_packet_extraction;
+pub mod pcap_extraction;
+pub mod pcap_struct;
+pub mod utils;
